Made the admin page look better and added QR functionality
I made it so that the admin page look better and works more easily, I also added a QR button that displays the QR of that artefact and you can download it in a  png, jpg, jpeg format. I still need to fix the date range filter for the quests and I still need to add the Download all selected QR codes button and then make the QR generation all go though Lambda.