This commit is for the user data that needs to be saved
This now correctly shows artefacts that are in the user's artefacts collected, and also the quests now save in the users completed quests array. I now need to make it that when a user scans an artefacts that it saves the artefact in their artefacts list as well as making the quest hints, completed, incorrect toasts.